# docs/swagger.yaml
definitions:
...
  models.Cluster:
    properties:
      login_nodes:
        items:
          $ref: '#/definitions/models.LoginNode'
        type: array
      name:
        type: string
    type: object
  models.LoginInfo:
    properties:
      login_node:
        $ref: '#/definitions/models.LoginNode'
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.LoginNode:
    properties:
      host:
        type: string
      name:
        type: string
      port:
        type: string
    type: object
  models.SacctRequest:
    properties:
      accounts:
        items:
          type: string
        type: array
      allclusters:
        type: boolean
      allusers:
        type: boolean
      arrayjobs:
        type: boolean
      brief:
        type: boolean
      clusters:
        items:
          type: string
        type: array
      completion:
        type: boolean
      duplicates:
        type: boolean
      endtime:
        type: string
      format:
        type: string
      jobids:
        items:
          type: string
        type: array
      jobnames:
        items:
          type: string
        type: array
      long:
        type: boolean
      maxcpus:
        type: integer
      maxnodes:
        type: integer
      mincpus:
        type: integer
      minnodes:
        type: integer
      nodelist:
        items:
          type: string
        type: array
      noheader:
        type: boolean
      parsable:
        type: boolean
      partitions:
        items:
          type: string
        type: array
      qos:
        items:
          type: string
        type: array
      starttime:
        type: string
      states:
        items:
          type: string
        type: array
      truncate:
        type: boolean
      users:
        items:
          type: string
        type: array
    type: object
...
  models.SqueueRequest:
    properties:
      accounts:
        items:
          type: string
        type: array
      all:
        type: boolean
      array:
        type: boolean
      clusters:
        items:
          type: string
        type: array
      federation:
        type: boolean
      format:
        type: string
      format_long:
        type: string
      hide:
        type: boolean
      host:
        type: string
      iterate:
        type: integer
      jobs:
        items:
          type: string
        type: array
      json:
        type: string
      licenses:
        items:
          type: string
        type: array
      local:
        type: boolean
      long:
        type: boolean
      names:
        items:
          type: string
        type: array
      noconvert:
        type: boolean
      nodelist:
        items:
          type: string
        type: array
      noheader:
        type: boolean
      only_job_state:
        type: boolean
      partitions:
        items:
          type: string
        type: array
      password:
        type: string
      port:
        type: integer
      privatekey:
        type: string
      qos:
        items:
          type: string
        type: array
      reservation:
        type: string
      sibling:
        type: boolean
      sort:
        items:
          type: string
        type: array
      start:
        type: boolean
      states:
        items:
          type: string
        type: array
      steps:
        items:
          type: string
        type: array
      username:
        type: string
      users:
        items:
          type: string
        type: array
      verbose:
        type: boolean
      yaml:
        type: string
    required:
    - host
    - username
    type: object
  models.User:
    properties:
      cluster:
        $ref: '#/definitions/models.Cluster'
      home_path:
        type: string
      password:
        type: string
      private_key:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/slurm/sacct/job/user/{user}/:
    post:
      consumes:
      - application/json
      description: 获取作业会计信息，包括作业ID、用户、状态、提交时间等详细信息
      parameters:
      - description: sacct请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SacctRequest'
      - description: SSH会话密钥
        example: '"tsh_a2e932b625c0d598db3800aa91b92016"'
        in: header
        name: sessionKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功，返回作业会计信息列表
          schema:
            properties:
              home_path:
                type: string
              session_key:
                type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: 认证失败，用户名或密码错误
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器内部错误或SSH连接失败
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 获取指定用户作业会计信息
      tags:
      - 作业管理
...
  /api/v1/slurm/squeue/jobs/:
    post:
      consumes:
      - application/json
      description: 获取作业列表信息，包括作业ID、用户、状态、提交时间等信息
      parameters:
      - description: squeue请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SqueueRequest'
      - description: SSH会话密钥
        example: '"tsh_a2e932b625c0d598db3800aa91b92016"'
        in: header
        name: sessionKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功，返回作业会计信息列表
          schema:
            properties:
              home_path:
                type: string
              session_key:
                type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: 认证失败，用户名或密码错误
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器内部错误或SSH连接失败
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 获取作业列表
      tags:
      - 作业管理
  /api/v1/user/login:
    post:
      consumes:
      - application/json
      description: 创建SSH连接会话，验证用户凭据并返回会话密钥用于后续API调用
      parameters:
      - description: 登录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginInfo'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功，返回用户主目录路径和会话密钥
          schema:
            properties:
              home_path:
                type: string
              session_key:
                type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: 认证失败，用户名或密码错误
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器内部错误或SSH连接失败
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 用户登录认证
      tags:
      - 认证管理
  /api/v1/user/logout/:
    get:
      consumes:
      - application/json
      description: 删除用户会话
      parameters:
      - description: 会话��钥
        in: header
        name: session_key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器内部错误或用户未登录
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 用户登出
      tags:
      - 认证管理
swagger: "2.0"